update exclusions: move to util/exclusions.js load from existing yaml file
